Split up plugin logic and move module transform out of Program enter
