--- packages/plugin/src/helpers/utils.js
@@ -1,3 +1,0 @@
-// export function equalsIgnoreCase() {
-//
-// }
