+++ packages/plugin/src/utils/ast.js
